requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
Flask-Login==0.6.3
ldap3==2.9.1
pywin32==311
